Fix: Fix various bugs in Lexer and Reader
Fix the bug that Lexer stuck in infinite loop when "0" is in the middle
of a string.

Fix the bug that Reader doesn't reset the value of _curr_line when
read_string is called more than once.

Fix the bug that Reader doesn't clear _lines (std::vector<line_info*>)
when read_string is called more than once.