Remove build system
After thinking this for a while, I decided to not implement a build
system (like zig or jai) for tan.

The reasons are:

- Extra work for maintaining the compiler
- Hard to integrate tan into other projects using a different build
system

Besides, using CLI to build stuff is easier to implement, and many tools
support it. Also, build system is not the main focus of tan, at least for now.